Update machine to smaller area.
This is the projector turned all the way over to the tightest zoom
level, which gives me a positive stop. I will need to monkey with this
again if I need a larger build area, but most of what I'm doing now is
small. Repeatable setting, greater light intensity, shorter cure
times. Let's use this for now.