Reduce layer times.
Trying to reduce the overcure I'm seeing by shortening the time. 6
seconds is still slightly overcured, but only slightly.

Lowered the base time too, as 30 seconds seems to provide as good
adhesion as 60. I'm not having issues with things not sticking to the
base plates, more that supports are breaking.