WIP committing green profile before switching back to red.
THIS DOES NOT WORK. At this point, I'm having pretty serious undercure
of most layers. When it does print, it looks overcured, but that could
be loss of detail due to other factors. mUVe3D support suggests that
my loss of fine detail might be because the green radiates some UV
when hardened, so I get loss of detail, and red is better in this regard.

Based on my testing so far, the best cure time for this resin at 100
micron is in the 6-8 second range, but it's best for solid/chunky
models. I will likely adjust this and use it up on pieces where detail
doesn't matter.