Restore faster lift rate and shorter lift.
For the first 5mm (100 layers @ .050 and 50 layers @ .100), use the
default 1mm/s lift (supplied in config). After that, go to 5mm/s
(because we should be out of the resin, so less vacuum). I may change
this to 10mm/s if 5mm/s works fine.

Also change to 2mm lift, because there seems to be no difference in
prints between the default 3.5mm and 2mm.